math: q_math becomes C++, and the move was not free
First mechanical step of section 40 - vec3_t as a type rather than an array -
and the step exists to be inert: the file changes language and nothing changes
behaviour. Inertness is a claim, so it is measured. The object this translation
unit produces was compiled both ways with the build's own flags and compared:
the .text section is byte-identical, and the only difference in the symbol
table is that two file-static objects, holdrand and bytedirs, get C++ internal
names. Linked, the four stripped binaries differ in exactly twenty bytes each,
all of them inside .note.gnu.build-id - which hashes the names of the linker's
inputs, and one of those names just changed from q_math.c.o to q_math.cpp.o.

That is the second measurement. The first one found a defect.

C++ was producing NormalToLatLong twenty-one bytes smaller than C, and the
reason was not optimisation. In C, atan2 and acos take doubles: a float
argument is promoted, the answer computed in double and narrowed back. In C++,
<math.h> brings the float overloads with it, so the same unchanged source
silently calls atan2f and acosf instead - different functions at a different
precision. Neither is correctly rounded, and the result is truncated to an int
two lines later, so a normal sitting on a boundary encodes to a different byte.
A behaviour change wearing a size change's clothes, in the one place out of
1433 lines where the language was visible at all.

The casts to double put it back. They are the only two edits in the file, and
the comment beside them says how they were found, because the next file to make
this move will have its own.

+++ shared/qcommon/q_math.cpp
@@ -182,10 +182,27 @@ void NormalToLatLong( const vec3_t normal, byte bytes[2] )
 	{
 		int	a, b;
 
-		a = (int)(RAD2DEG( (float)atan2( normal[1], normal[0] ) ) * (255.0f / 360.0f ));
+		// The casts to double are what keeps this file's move to C++ inert, and
+		// they are the only two in 1433 lines.
+		//
+		// In C, atan2 and acos take doubles: a float argument is promoted, the
+		// answer is computed in double and narrowed back. In C++, <math.h>
+		// brings in the float overloads too, so the same source silently calls
+		// atan2f and acosf instead - a different function, computed at a
+		// different precision. Neither of those two is correctly rounded, so
+		// the answers differ in the last bits, and the last bits are then
+		// truncated to an integer by the cast below: a normal sitting on a
+		// boundary can encode to a different byte.
+		//
+		// Found by measurement rather than by reading. The object file this
+		// translation unit produces is byte-identical between the two
+		// languages except for these two calls, which cost this function
+		// twenty-one bytes of code; that was the only difference in the whole
+		// file, and it was a behaviour change wearing a size change's clothes.
+		a = (int)(RAD2DEG( (float)atan2( (double)normal[1], (double)normal[0] ) ) * (255.0f / 360.0f ));
 		a &= 0xff;
 
-		b = (int)(RAD2DEG( (float)acos( normal[2] ) ) * ( 255.0f / 360.0f ));
+		b = (int)(RAD2DEG( (float)acos( (double)normal[2] ) ) * ( 255.0f / 360.0f ));
 		b &= 0xff;
 
 		bytes[0] = b;	// longitude
